update wishlist query-option location
